refactor: Smaller refactoring
